feat: expose ava before/after hooks through testProp
In the case where the user supplies his own `ExecutionContext` as well
as before/after hooks, this means the user only needs to set the
`ExecutionContext` with a type assertion one time (on `testProp`) rather
than on `testProp` and on the AVA `test`. For example:

```typescript
import { testProp as anyTestProp, PropertyTestInterface } from 'ava-fast-check';

type TestContext = {
    myVariable: string
};

const testProp = anyTestProp as PropertyTestInterface<TestContext>;

testProp.before(t => {
    // t.context.myVariable is typed as `string` here
});

testProp(
    'should do something',
    [fc.string()],
    (t, received) => {
        // t.context.myVariable is also typed as `string` here
    }
);
```

Formerly, the user had to use a type assertion on both `test` and
`testProp` to have this static typing in both hooks and property test
bodies.