Prevent test.sh from failing
package.json declares a package of type commonjs, thus it cannot use import. In the past ava was not checking this constraint so it worked, now that it checks it require has to be used for this test (import is used by other tests not depending of a package of type commonjs).